+++ packages/vue-composable/src/api.ts
@@ -1,3 +1,5 @@
+/* istanbul ignore file */
+
 export {
   ref,
   isRef,