+++ packages/vue-composable/src/date/performanceNow.ts
@@ -1,11 +1,11 @@
-import { useNow, NowOptions } from "./now";
+import { useNow, NowOptions, NowReturn } from "./now";
 import { isBoolean } from "../utils";
 
 /**
  * Reactive `performance.now()` with specified refresh rate
  * @param options - Configuration
  */
-export function usePerformanceNow(options?: NowOptions) {
+export function usePerformanceNow(options?: NowOptions): NowReturn {
   const refreshMs = (options && options.refreshMs) || 1000;
   const sync = options && isBoolean(options.sync) ? options.sync : true;