+++ packages/vue-composable/src/index.ts
@@ -12,3 +12,7 @@ export * from "./date";
 export * from "./format";
 export * from "./validation";
 export * from "./i18n";
+
+export const VERSION = __VERSION__;
+export const VUE_VERSION: "2" | "3" = __VUE_2__ ? "2" : "3";
+export const COMMIT = __COMMIT__;